change/unify ball trail strength, global difficulty level and playfield reflection strength to range 0..100
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2501 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ pintable.h
@@ -716,9 +716,9 @@ public:
    float m_AOScale;
 
    int   m_useReflectionForBalls;
-   int   m_playfieldReflectionStrength;
+   float m_playfieldReflectionStrength;
    int   m_useTrailForBalls;
-   int   m_ballTrailStrength;
+   float m_ballTrailStrength;
    float m_ballPlayfieldReflectionStrength;
    int   m_useAA;
    int   m_useFXAA;
